simple-rag-chatbot/pyproject.toml
[project]
name = "simple-rag-chatbot"
version = "0.1.0"
description = "Simple RAG chatbot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ollama>=0.5.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "surrealdb>=1.0.4",
]
